+++ src/area-chart/__tests__/area-chart-async-store.test.ts
@@ -1,6 +1,6 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
-import AsyncStore, { useReaction, useSelector } from '../model/async-store';
+import AsyncStore, { useReaction, useSelector } from '../../internal/utils/async-store';
 import { renderHook, act } from '../../__tests__/render-hook';
 
 describe('AreaChart AsyncStore', () => {
