+++ src/area-chart/elements/area-chart-legend.tsx
@@ -4,7 +4,7 @@ import React, { memo, useMemo } from 'react';
 
 import { AreaChartProps } from '../interfaces';
 import ChartLegend from '../../internal/components/chart-legend';
-import { useSelector } from '../model/async-store';
+import { useSelector } from '../../internal/utils/async-store';
 import { ChartModel } from '../model';
 
 export default memo(AreaChartLegend) as typeof AreaChartLegend;
