+++ src/area-chart/model/interactions-store.ts
@@ -3,7 +3,7 @@
 import { ChartModel } from './index';
 import { AreaChartProps } from '../interfaces';
 
-import AsyncStore from './async-store';
+import AsyncStore from '../../internal/utils/async-store';
 
 const initialState = Object.freeze({
   highlightedX: null,
