+++ src/area-chart/model/use-chart-model.ts
@@ -7,7 +7,7 @@ import { findClosest, circleIndex } from './utils';
 import { nodeContains } from '../../internal/utils/dom';
 import { KeyCode } from '../../internal/keycode';
 import { XDomain, XScaleType, YDomain, YScaleType } from '../../internal/components/cartesian-chart/interfaces';
-import { useReaction } from './async-store';
+import { useReaction } from '../../internal/utils/async-store';
 import computeChartProps from './compute-chart-props';
 import createSeriesDecorator from './create-series-decorator';
 import InteractionsStore from './interactions-store';
